moved monitor controller to controllers module
