+++ examples/digital_fingerprinting/production/morpheus/dfp/stages/dfp_mlflow_model_writer.py
@@ -20,10 +20,10 @@ import mrc
 from mrc.core import operators as ops
 
 from morpheus.config import Config
+from morpheus.controllers.mlflow_model_writer_controller import MLFlowModelWriterController
 from morpheus.messages.multi_ae_message import MultiAEMessage
 from morpheus.pipeline.single_port_stage import SinglePortStage
 from morpheus.pipeline.stream_pair import StreamPair
-from morpheus.utils.controllers.mlflow_model_writer_controller import MLFlowModelWriterController
 
 # Setup conda environment
 conda_env = {
