+++ morpheus/stages/general/monitor_stage.py
@@ -21,10 +21,10 @@ from tqdm import tqdm
 
 from morpheus.cli.register_stage import register_stage
 from morpheus.config import Config
+from morpheus.controllers.monitor_controller import MonitorController
 from morpheus.pipeline.single_port_stage import SinglePortStage
 from morpheus.pipeline.stream_pair import StreamPair
 from morpheus.utils.logger import LogLevels
-from morpheus.utils.monitor_utils import MonitorController
 
 logger = logging.getLogger(__name__)
