+++ tests/test_monitor_stage.py
@@ -59,7 +59,7 @@ def test_constructor(config):
     assert stage._mc._determine_count_fn is two_x
 
 
-@mock.patch('morpheus.utils.monitor_utils.MorpheusTqdm')
+@mock.patch('morpheus.controllers.monitor_controller.MorpheusTqdm')
 def test_on_start(mock_morph_tqdm, config):
     mock_morph_tqdm.return_value = mock_morph_tqdm
 
@@ -72,7 +72,7 @@ def test_on_start(mock_morph_tqdm, config):
     assert stage._mc._progress is mock_morph_tqdm
 
 
-@mock.patch('morpheus.utils.monitor_utils.MorpheusTqdm')
+@mock.patch('morpheus.controllers.monitor_controller.MorpheusTqdm')
 def test_stop(mock_morph_tqdm, config):
     mock_morph_tqdm.return_value = mock_morph_tqdm
 
@@ -88,7 +88,7 @@ def test_stop(mock_morph_tqdm, config):
     mock_morph_tqdm.close.assert_called_once()
 
 
-@mock.patch('morpheus.utils.monitor_utils.MorpheusTqdm')
+@mock.patch('morpheus.controllers.monitor_controller.MorpheusTqdm')
 def test_refresh(mock_morph_tqdm, config):
     mock_morph_tqdm.return_value = mock_morph_tqdm
 
@@ -134,7 +134,7 @@ def test_auto_count_fn_not_impl(config, value: typing.Any):
         stage._mc.auto_count_fn(value)
 
 
-@mock.patch('morpheus.utils.monitor_utils.MorpheusTqdm')
+@mock.patch('morpheus.controllers.monitor_controller.MorpheusTqdm')
 def test_progress_sink(mock_morph_tqdm, config):
     mock_morph_tqdm.return_value = mock_morph_tqdm
